requirements.txt
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.7.0
requests==2.31.0
notion-client==2.0.0
